Redesign site

+++ build.sh
@@ -3,16 +3,23 @@ set -e
 source ./venv/bin/activate
 rm -rf ./site/*
 
-# Create history index
-echo Creating history index...
-rm -f history.csv
-find notes -name "*.md" | while read line; do
-    echo -n "\"$line\"," >> history.csv
-    git log --pretty=format:"%ad," --date=short --diff-filter=A -- "$line" | tail -n1  >> history.csv
-    git log -1 --pretty=format:"%ad" --date=short -- "$line" | tail -n1  >> history.csv
-    echo "" >> history.csv
-done
-echo History index created!
+
+# skip creating history index if -f flag is passed
+if [ "$1" == "-f" ]; then
+    echo "Skipping history index creation"
+    shift
+else
+    # Create history index
+    echo Creating history index...
+    rm -f history.csv
+    find notes -name "*.md" | while read line; do
+        echo -n "\"$line\"," >> history.csv
+        git log --pretty=format:"%ad," --date=short --diff-filter=A -- "$line" | tail -n1  >> history.csv
+        git log -1 --pretty=format:"%ad" --date=short -- "$line" | tail -n1  >> history.csv
+        echo "" >> history.csv
+    done
+    echo History index created!
+fi
 
 echo Activated venv
 echo Generating...
